Split jvm_binary.py into jvm_binary.py and jvm_app.py.
- Moved test_jvm_app.py into the tests/python/pants_test/backend/jvm/targets directory
- Added some explicit tests for the JvmBinary target.
- Minor fix to jvm_binary.py to raise an error about the 'source' attribute before
  it gets passed to the super() constructor.
