+++ src/python/pants/backend/project_info/tasks/dependencies.py
@@ -8,7 +8,7 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 from twitter.common.collections import OrderedSet
 
 from pants.backend.core.tasks.console_task import ConsoleTask
-from pants.backend.jvm.targets.jvm_binary import JvmApp
+from pants.backend.jvm.targets.jvm_app import JvmApp
 from pants.base.exceptions import TaskError
 from pants.base.payload_field import JarsField, PythonRequirementsField
