+++ tests/python/pants_test/backend/project_info/tasks/BUILD
@@ -67,13 +67,11 @@ python_tests(
   name = 'filedeps',
   sources = ['test_filedeps.py'],
   dependencies = [
-    'src/python/pants/backend/codegen/targets:java',
-    'src/python/pants/backend/core/targets:common',
+    'src/python/pants/backend/codegen:plugin',
+    'src/python/pants/backend/core:plugin',
+    'src/python/pants/backend/jvm:plugin',
     'src/python/pants/backend/jvm/targets:java',
-    'src/python/pants/backend/jvm/targets:jvm',
-    'src/python/pants/backend/jvm/targets:scala',
     'src/python/pants/backend/project_info/tasks:filedeps',
-    'src/python/pants/base:build_file_aliases',
     'src/python/pants/base:config',
     'tests/python/pants_test/tasks:base',
   ],
